fix(clippy): satisfy clippy 1.97 lints; pin CI toolchain to 1.97.0
CI's dtolnay/rust-toolchain@stable had advanced to rustc 1.97.0, whose clippy
adds/tightens lints my local (stale mise 1.94) never saw — the -clippy gate
failed on ~65 sites. Installed 1.97.0 locally and cleaned the workspace against
it: cargo +1.97.0 clippy --fix for the machine-applicable majority, then scoped
#[allow(...)] with justifications for the intentional patterns the new lints flag:
- manual_checked_ops: explicit `if divisor == 0` divide-by-zero guards (clearer
  than checked_div in this literate-programming repo, and documented per fn),
- while_let_loop / explicit_counter_loop: explicit loops with internal break
  conditions / meaningful counters,
- collapsible_match, unnecessary_sort_by (descending comparators), question_mark:
  kept the clearer form.

Pin the CI toolchain to dtolnay/rust-toolchain@1.97.0 (was @stable). A blocking
clippy gate on an unpinned @stable would spontaneously red main on every Rust
release with zero code change; pinning makes the gate deterministic. Bumping is
deliberate (change the version + re-run clippy --fix in the same PR). lessons.md
updated with the pin rationale and the local/CI clippy-version-match requirement.

Verified: cargo +1.97.0 clippy --workspace --all-targets -- -D warnings is clean
(excluding host-incompatible paint-vm-*/os-kernel).

+++ code/packages/rust/micro-qr/src/lib.rs
@@ -903,7 +903,7 @@ pub fn encode(
     let mut bits: Vec<bool> = Vec::new();
     for (cw_idx, &cw) in final_cw.iter().enumerate() {
         let bits_in_cw = if cfg.m1_half_cw && cw_idx == cfg.data_cw - 1 { 4 } else { 8 };
-        for b in (0..bits_in_cw as u32).rev() {
+        for b in (0..bits_in_cw).rev() {
             bits.push(((cw >> (b + (8 - bits_in_cw))) & 1) == 1);
         }
     }
